requirements.txt
anyio==3.6.1
asyncpg==0.26.0
click==8.1.3
databases==0.6.1
ecdsa==0.18.0
fastapi==0.80.0
fastapi_auth_middleware==1.0.2
greenlet==1.1.3
h11==0.13.0
idna==3.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pydantic==1.9.2
python-dateutil==2.8.2
python-jose==3.3.0
pytimeparse==1.1.8
rsa==4.9
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
SQLAlchemy==1.4.40
SQLAlchemy-serializer==1.4.1
starlette==0.19.1
timedelta==2020.12.3
typing_extensions==4.3.0
uvicorn==0.18.3
